--- projects/ontimize-web-ngx-report/src/lib/o-components.ts
@@ -1,7 +1,0 @@
-import { OReportRoutingModule } from './components/report/o-report-routing.module';
-
-export * from './components/report/o-report-home/o-report-home.component';
-
-export const OREPORT_MODULES: any = [
-  OReportRoutingModule
-];
